Login Screen Functionality
Login screen functionality has been added and
tested. The layout needs work, and the user experience
could be better, but this is only a start. A basis has
been build for the purpose of authorization and authentication.

Proxy server has also been added, because a CORS error crept
up in the browser. This Fixed the CORS error by using a single
host (ie, the 4200 port) instead of two (ie, also port 8081).